+++ quackosm/functions.py
@@ -9,7 +9,7 @@ from pathlib import Path
 from typing import Any, Literal, Optional, Union
 
 import geopandas as gpd
-from pandas.util._decorators import deprecate, deprecate_kwarg
+from rq_geo_toolkit.duckdb import DuckDBConnKwargs
 from shapely.geometry.base import BaseGeometry
 
 from quackosm._constants import (
@@ -18,6 +18,7 @@ from quackosm._constants import (
     PARQUET_ROW_GROUP_SIZE,
     PARQUET_VERSION,
 )
+from quackosm._deprecate import deprecate, deprecate_kwarg
 from quackosm._osm_tags_filters import GroupedOsmTagsFilter, OsmTagsFilter
 from quackosm._osm_way_polygon_features import OsmWayPolygonConfig
 from quackosm._rich_progress import VERBOSITY_MODE
@@ -60,6 +61,7 @@ def convert_pbf_to_duckdb(
     debug_memory: bool = False,
     debug_times: bool = False,
     cpu_limit: Optional[int] = None,
+    duckdb_conn_kwargs: Optional[DuckDBConnKwargs] = None,
 ) -> Path:
     """
     Convert PBF file to DuckDB file.
@@ -130,6 +132,8 @@ def convert_pbf_to_duckdb(
             step has been executed. Defaults to `False`.
         cpu_limit (int, optional): Max number of threads available for processing.
             If `None`, will use all available threads. Defaults to `None`.
+        duckdb_conn_kwargs (Optional[DuckDBConnKwargs], optional): Additional kwargs used to
+            provision a duckdb connection. Defaults to None.
 
     Returns:
         Path: Path to the generated DuckDB file.
@@ -287,6 +291,7 @@ def convert_pbf_to_duckdb(
         debug_memory=debug_memory,
         debug_times=debug_times,
         cpu_limit=cpu_limit,
+        duckdb_conn_kwargs=duckdb_conn_kwargs,
     ).convert_pbf_to_duckdb(
         pbf_path=pbf_path,
         result_file_path=result_file_path,
@@ -325,6 +330,7 @@ def convert_geometry_to_duckdb(
     debug_memory: bool = False,
     debug_times: bool = False,
     cpu_limit: Optional[int] = None,
+    duckdb_conn_kwargs: Optional[DuckDBConnKwargs] = None,
 ) -> Path:
     """
     Get a DuckDB file with OpenStreetMap features within given geometry.
@@ -406,6 +412,8 @@ def convert_geometry_to_duckdb(
             step has been executed. Defaults to `False`.
         cpu_limit (int, optional): Max number of threads available for processing.
             If `None`, will use all available threads. Defaults to `None`.
+        duckdb_conn_kwargs (Optional[DuckDBConnKwargs], optional): Additional kwargs used to
+            provision a duckdb connection. Defaults to None.
 
     Returns:
         Path: Path to the generated DuckDB file.
@@ -521,6 +529,7 @@ def convert_geometry_to_duckdb(
         debug_memory=debug_memory,
         debug_times=debug_times,
         cpu_limit=cpu_limit,
+        duckdb_conn_kwargs=duckdb_conn_kwargs,
     ).convert_geometry_to_duckdb(
         result_file_path=result_file_path,
         keep_all_tags=keep_all_tags,
@@ -557,6 +566,7 @@ def convert_osm_extract_to_duckdb(
     debug_memory: bool = False,
     debug_times: bool = False,
     cpu_limit: Optional[int] = None,
+    duckdb_conn_kwargs: Optional[DuckDBConnKwargs] = None,
 ) -> Path:
     """
     Get a single OpenStreetMap extract from a given source and transform it to a DuckDB file.
@@ -630,6 +640,8 @@ def convert_osm_extract_to_duckdb(
             step has been executed. Defaults to `False`.
         cpu_limit (int, optional): Max number of threads available for processing.
             If `None`, will use all available threads. Defaults to `None`.
+        duckdb_conn_kwargs (Optional[DuckDBConnKwargs], optional): Additional kwargs used to
+            provision a duckdb connection. Defaults to None.
 
     Returns:
         Path: Path to the generated DuckDB file.
@@ -706,6 +718,7 @@ def convert_osm_extract_to_duckdb(
         debug_memory=debug_memory,
         debug_times=debug_times,
         cpu_limit=cpu_limit,
+        duckdb_conn_kwargs=duckdb_conn_kwargs,
     ).convert_pbf_to_duckdb(
         pbf_path=downloaded_osm_extract,
         result_file_path=result_file_path,
@@ -742,6 +755,7 @@ def convert_pbf_to_parquet(
     debug_memory: bool = False,
     debug_times: bool = False,
     cpu_limit: Optional[int] = None,
+    duckdb_conn_kwargs: Optional[DuckDBConnKwargs] = None,
 ) -> Path:
     """
     Convert PBF file to GeoParquet file.
@@ -814,6 +828,8 @@ def convert_pbf_to_parquet(
             step has been executed. Defaults to `False`.
         cpu_limit (int, optional): Max number of threads available for processing.
             If `None`, will use all available threads. Defaults to `None`.
+        duckdb_conn_kwargs (Optional[DuckDBConnKwargs], optional): Additional kwargs used to
+            provision a duckdb connection. Defaults to None.
 
     Returns:
         Path: Path to the generated GeoParquet file.
@@ -969,6 +985,7 @@ def convert_pbf_to_parquet(
         debug_memory=debug_memory,
         debug_times=debug_times,
         cpu_limit=cpu_limit,
+        duckdb_conn_kwargs=duckdb_conn_kwargs,
     ).convert_pbf_to_parquet(
         pbf_path=pbf_path,
         result_file_path=result_file_path,
@@ -1007,6 +1024,7 @@ def convert_geometry_to_parquet(
     debug_memory: bool = False,
     debug_times: bool = False,
     cpu_limit: Optional[int] = None,
+    duckdb_conn_kwargs: Optional[DuckDBConnKwargs] = None,
 ) -> Path:
     """
     Get a GeoParquet file with OpenStreetMap features within given geometry.
@@ -1090,6 +1108,8 @@ def convert_geometry_to_parquet(
             step has been executed. Defaults to `False`.
         cpu_limit (int, optional): Max number of threads available for processing.
             If `None`, will use all available threads. Defaults to `None`.
+        duckdb_conn_kwargs (Optional[DuckDBConnKwargs], optional): Additional kwargs used to
+            provision a duckdb connection. Defaults to None.
 
     Returns:
         Path: Path to the generated GeoParquet file.
@@ -1202,6 +1222,7 @@ def convert_geometry_to_parquet(
         debug_memory=debug_memory,
         debug_times=debug_times,
         cpu_limit=cpu_limit,
+        duckdb_conn_kwargs=duckdb_conn_kwargs,
     ).convert_geometry_to_parquet(
         result_file_path=result_file_path,
         keep_all_tags=keep_all_tags,
@@ -1238,6 +1259,7 @@ def convert_osm_extract_to_parquet(
     debug_memory: bool = False,
     debug_times: bool = False,
     cpu_limit: Optional[int] = None,
+    duckdb_conn_kwargs: Optional[DuckDBConnKwargs] = None,
 ) -> Path:
     """
     Get a single OpenStreetMap extract from a given source and transform it to a GeoParquet file.
@@ -1313,6 +1335,8 @@ def convert_osm_extract_to_parquet(
             step has been executed. Defaults to `False`.
         cpu_limit (int, optional): Max number of threads available for processing.
             If `None`, will use all available threads. Defaults to `None`.
+        duckdb_conn_kwargs (Optional[DuckDBConnKwargs], optional): Additional kwargs used to
+            provision a duckdb connection. Defaults to None.
 
     Returns:
         Path: Path to the generated GeoParquet file.
@@ -1388,6 +1412,7 @@ def convert_osm_extract_to_parquet(
         debug_memory=debug_memory,
         debug_times=debug_times,
         cpu_limit=cpu_limit,
+        duckdb_conn_kwargs=duckdb_conn_kwargs,
     ).convert_pbf_to_parquet(
         pbf_path=downloaded_osm_extract,
         result_file_path=result_file_path,
@@ -1401,7 +1426,7 @@ def convert_osm_extract_to_parquet(
     return Path(result_path)
 
 
-@deprecate_kwarg(old_arg_name="file_paths", new_arg_name="pbf_path")  # type: ignore
+@deprecate_kwarg(old_arg_name="file_paths", new_arg_name="pbf_path")
 def convert_pbf_to_geodataframe(
     pbf_path: Union[str, Path, Iterable[Union[str, Path]]],
     tags_filter: Optional[Union[OsmTagsFilter, GroupedOsmTagsFilter]] = None,
@@ -1423,6 +1448,7 @@ def convert_pbf_to_geodataframe(
     debug_memory: bool = False,
     debug_times: bool = False,
     cpu_limit: Optional[int] = None,
+    duckdb_conn_kwargs: Optional[DuckDBConnKwargs] = None,
 ) -> gpd.GeoDataFrame:
     """
     Get features GeoDataFrame from a PBF file or list of PBF files.
@@ -1492,6 +1518,8 @@ def convert_pbf_to_geodataframe(
             step has been executed. Defaults to `False`.
         cpu_limit (int, optional): Max number of threads available for processing.
             If `None`, will use all available threads. Defaults to `None`.
+        duckdb_conn_kwargs (Optional[DuckDBConnKwargs], optional): Additional kwargs used to
+            provision a duckdb connection. Defaults to None.
 
     Returns:
         gpd.GeoDataFrame: GeoDataFrame with OSM features.
@@ -1624,6 +1652,7 @@ def convert_pbf_to_geodataframe(
         debug_memory=debug_memory,
         debug_times=debug_times,
         cpu_limit=cpu_limit,
+        duckdb_conn_kwargs=duckdb_conn_kwargs,
     ).convert_pbf_to_geodataframe(
         pbf_path=pbf_path,
         keep_all_tags=keep_all_tags,
@@ -1657,6 +1686,7 @@ def convert_geometry_to_geodataframe(
     debug_memory: bool = False,
     debug_times: bool = False,
     cpu_limit: Optional[int] = None,
+    duckdb_conn_kwargs: Optional[DuckDBConnKwargs] = None,
 ) -> gpd.GeoDataFrame:
     """
     Get features GeoDataFrame with OpenStreetMap features within given geometry.
@@ -1734,6 +1764,8 @@ def convert_geometry_to_geodataframe(
             step has been executed. Defaults to `False`.
         cpu_limit (int, optional): Max number of threads available for processing.
             If `None`, will use all available threads. Defaults to `None`.
+        duckdb_conn_kwargs (Optional[DuckDBConnKwargs], optional): Additional kwargs used to
+            provision a duckdb connection. Defaults to None.
 
     Returns:
         gpd.GeoDataFrame: GeoDataFrame with OSM features.
@@ -1806,6 +1838,7 @@ def convert_geometry_to_geodataframe(
         debug_memory=debug_memory,
         debug_times=debug_times,
         cpu_limit=cpu_limit,
+        duckdb_conn_kwargs=duckdb_conn_kwargs,
     ).convert_geometry_to_geodataframe(
         keep_all_tags=keep_all_tags,
         explode_tags=explode_tags,
@@ -1837,6 +1870,7 @@ def convert_osm_extract_to_geodataframe(
     debug_memory: bool = False,
     debug_times: bool = False,
     cpu_limit: Optional[int] = None,
+    duckdb_conn_kwargs: Optional[DuckDBConnKwargs] = None,
 ) -> gpd.GeoDataFrame:
     """
     Get a single OpenStreetMap extract from a given source and return it as a GeoDataFrame.
@@ -1906,6 +1940,8 @@ def convert_osm_extract_to_geodataframe(
             step has been executed. Defaults to `False`.
         cpu_limit (int, optional): Max number of threads available for processing.
             If `None`, will use all available threads. Defaults to `None`.
+        duckdb_conn_kwargs (Optional[DuckDBConnKwargs], optional): Additional kwargs used to
+            provision a duckdb connection. Defaults to None.
 
     Returns:
         gpd.GeoDataFrame: GeoDataFrame with OSM features.
@@ -1974,6 +2010,7 @@ def convert_osm_extract_to_geodataframe(
         debug_memory=debug_memory,
         debug_times=debug_times,
         cpu_limit=cpu_limit,
+        duckdb_conn_kwargs=duckdb_conn_kwargs,
     ).convert_pbf_to_geodataframe(
         pbf_path=downloaded_osm_extract,
         keep_all_tags=keep_all_tags,
@@ -1985,29 +2022,29 @@ def convert_osm_extract_to_geodataframe(
 
 
 convert_pbf_to_gpq = deprecate(
-    "convert_pbf_to_gpq",
-    convert_pbf_to_parquet,
-    "0.8.1",
+    name="convert_pbf_to_gpq",
+    alternative=convert_pbf_to_parquet,
+    version="0.8.1",
     msg="Use `convert_pbf_to_parquet` instead. Deprecated since 0.8.1 version.",
 )
 
 convert_geometry_to_gpq = deprecate(
-    "convert_geometry_to_gpq",
-    convert_geometry_to_parquet,
-    "0.8.1",
+    name="convert_geometry_to_gpq",
+    alternative=convert_geometry_to_parquet,
+    version="0.8.1",
     msg="Use `convert_geometry_to_parquet` instead. Deprecated since 0.8.1 version.",
 )
 
 get_features_gdf = deprecate(
-    "get_features_gdf",
-    convert_pbf_to_geodataframe,
-    "0.8.1",
+    name="get_features_gdf",
+    alternative=convert_pbf_to_geodataframe,
+    version="0.8.1",
     msg="Use `convert_pbf_to_geodataframe` instead. Deprecated since 0.8.1 version.",
 )
 
 get_features_gdf_from_geometry = deprecate(
-    "get_features_gdf_from_geometry",
-    convert_geometry_to_geodataframe,
-    "0.8.1",
+    name="get_features_gdf_from_geometry",
+    alternative=convert_geometry_to_geodataframe,
+    version="0.8.1",
     msg="Use `convert_geometry_to_geodataframe` instead. Deprecated since 0.8.1 version.",
 )